backend/Sentinel (2)/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
pymongo==4.6.0
pymongo[srv]==4.6.0
python-decouple==3.8
Pillow==10.1.0
python-dotenv==1.0.0
dnspython==2.4.2
